% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dozer.R
\name{clustering_difference_network}
\alias{clustering_difference_network}
\title{Conduct hierarchical clustering for "difference network" between diagnosis group.}
\usage{
clustering_difference_network(network1, network2, minClusterSize = 20)
}
\arguments{
\item{network1}{network in diagnosis group 1}

\item{network2}{network in diagnosis group 2}
}
\value{
cluster labels
}
\description{
Conduct hierarchical clustering for "difference network" between diagnosis group.
}
