% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dozer.R
\name{compute_centrality}
\alias{compute_centrality}
\title{Compute gene centrality in the hard-threshold gene co-expression network.}
\usage{
compute_centrality(network, threshold = 0.99)
}
\arguments{
\item{network}{Gene-gene correlation matrix.}

\item{threshold}{quantile threshold on the absolute correlations for hard-thresholding.}
}
\value{
list of gene centrality vectors, including degree, pagerank, betweenness and eigenvector centrality.
}
\description{
Compute gene centrality in the hard-threshold gene co-expression network.
}
