% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dozer.R
\name{diagnoistic_plot_cell_size}
\alias{diagnoistic_plot_cell_size}
\title{Generate a figure showing the slope of gene counts over trimmed total UMI counts.
If all gene slopes are centered around 1, using trimmed total UMI counts as adjustment for cell size is adequate.}
\usage{
diagnoistic_plot_cell_size(
  data,
  normalized_data = NULL,
  gene_group_quantile = NULL,
  n = 10
)
}
\arguments{
\item{data}{counts matrix}

\item{n}{number of bins for genes}
}
\value{
density plot of linear regression slopes of gene expression over trimmed total UMI counts.
}
\description{
Generate a figure showing the slope of gene counts over trimmed total UMI counts.
If all gene slopes are centered around 1, using trimmed total UMI counts as adjustment for cell size is adequate.
}
