% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dozer.R
\name{simulate_counts}
\alias{simulate_counts}
\title{Simulate gene expression matrix from a Gamma-Poisson distribution.}
\usage{
simulate_counts(corr, shape, scale, mlog, sdlog, ncell)
}
\arguments{
\item{corr}{Gene-gene correlation matrix served as ground truth.}

\item{shape, scale}{Gamma shape and scale parameter in the Gamma-Poisson distribution.}

\item{mlog, sdlog}{Mean and standard deviation of the log normal distribution for library size.}

\item{ncell}{Number of cells to be simulated.}
}
\value{
simulated gene expression matrix using Gamma Poisson distribution.
}
\description{
Simulate gene expression matrix from a Gamma-Poisson distribution.
}
