% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dozer.R
\name{trimmed_total_counts}
\alias{trimmed_total_counts}
\title{Compute trimmed total counts}
\usage{
trimmed_total_counts(data, max_weight = 1/300, delta = NULL)
}
\arguments{
\item{data}{counts matrix}

\item{max_weight}{genes with weights higher than max_weight will be trimmed}

\item{delta}{A threshold on the minimum total counts over median total counts, in order not to have extremely small trimmed total counts, which aims to stablize normalized counts.}
}
\value{
trimmed total counts per cell
}
\description{
Compute trimmed total counts
}
