% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dozer.R
\name{compute_gene_correlation}
\alias{compute_gene_correlation}
\title{Compute gene-gene correlation matrix}
\usage{
compute_gene_correlation(data, lib_size = NULL, covs = NULL)
}
\arguments{
\item{data}{A gene expression matrix, with rows representing genes and columns representing cells.}

\item{lib_size}{A vector of cell sequencing depth (optional).}

\item{covs}{A data frame or a list of other covariates for cells, e.g. batch labels.}
}
\value{
co-expression matrix in the "network" slot and gene noise to signal ratio vector in the "ratio" slot.
}
\description{
This function take gene expression matrix (gene by cell), cell sequencing depth
and other cell level covariates that need to be accounted for in expression normalization as input to
compute gene-gene correlation matrix.
}
